Fixed omp problem on windows.
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@7419 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ recognition/include/pcl/recognition/impl/hv/hv_go.hpp
@@ -292,7 +292,7 @@ void pcl::GlobalHypothesesVerification<ModelT, SceneT>::initialize()
   {
     pcl::ScopeTime tcues ("Computing clutter cues");
 #pragma omp parallel for schedule(dynamic, 4) num_threads(omp_get_num_procs())
-    for (size_t j = 0; j < recognition_models_.size (); j++)
+    for (int j = 0; j < static_cast<float> (recognition_models_.size ()); j++)
       computeClutterCue (recognition_models_[j]);
   }
